feat: add `--stdout` flag
This patch adds a `--stdout` (or `-j`...not convinced this is the best short-flag; suggestions welcome) flag to the `axe-cli` program. When this flag is provided, the "raw" result object from axe-core will be outputted to STDOUT. No other logs will be emitted on the STDOUT stream, as it's expected this stream will be piped into a file or some JSON-parsing tool.

Additionally all errors are being sent to STDERR (where they belong) rather than STDOUT. This enables us to keep error messages around even in `--stdout` mode.

Closes #15